+++ src/plugin.rs
@@ -3,16 +3,17 @@ use std::path::Path;
 use std::path::PathBuf;
 
 use anyhow::Result;
-use dprint_core::configuration::get_unknown_property_diagnostics;
 use dprint_core::configuration::ConfigKeyMap;
 use dprint_core::configuration::GlobalConfiguration;
 use dprint_core::configuration::ResolveConfigurationResult;
 use dprint_core::plugins::PluginHandler;
 use dprint_core::plugins::PluginInfo;
 
+use crate::configuration::Configuration;
 use crate::parser::parse_file;
 use crate::parser::Block;
 use crate::parser::Section;
+use crate::parser::StartTag;
 
 pub struct VuePluginHandler;
 
@@ -32,7 +33,7 @@ impl VuePluginHandler {
     }
 }
 
-impl PluginHandler<()> for VuePluginHandler {
+impl PluginHandler<Configuration> for VuePluginHandler {
     fn get_plugin_info(&mut self) -> PluginInfo {
         PluginInfo {
             name: env!("CARGO_PKG_NAME").to_string(),
@@ -52,19 +53,16 @@ impl PluginHandler<()> for VuePluginHandler {
     fn resolve_config(
         &mut self,
         config: ConfigKeyMap,
-        _global_config: &GlobalConfiguration,
-    ) -> ResolveConfigurationResult<()> {
-        ResolveConfigurationResult {
-            config: (),
-            diagnostics: get_unknown_property_diagnostics(config),
-        }
+        global_config: &GlobalConfiguration,
+    ) -> ResolveConfigurationResult<Configuration> {
+        Configuration::resolve(config, global_config)
     }
 
     fn format_text(
         &mut self,
         _file_path: &Path,
         file_text: &str,
-        _config: &(),
+        config: &Configuration,
         mut format_with_host: impl FnMut(&Path, String, &ConfigKeyMap) -> Result<String>,
     ) -> Result<String> {
         let mut buffer = String::new();
@@ -75,7 +73,7 @@ impl PluginHandler<()> for VuePluginHandler {
             match section {
                 Section::Root(value) => buffer.push_str(value),
                 Section::Block(Block {
-                    start_tag,
+                    start_tag: StartTag { name, lang },
                     content,
                     raw_start_tag,
                     raw_end_tag,
@@ -83,18 +81,40 @@ impl PluginHandler<()> for VuePluginHandler {
                     buffer.push_str(raw_start_tag);
                     buffer.push('\n');
 
-                    let lang = start_tag
-                        .lang
-                        .or_else(|| Self::default_lang(start_tag.name));
+                    let lang = lang.or_else(|| Self::default_lang(name));
 
                     if let Some(lang) = lang {
                         let file_path = PathBuf::from(format!("file.{}", lang));
 
-                        buffer.push_str(&format_with_host(
-                            &file_path,
-                            String::from(content),
-                            &HashMap::new(),
-                        )?);
+                        let pretty = {
+                            let pretty = format_with_host(
+                                &file_path,
+                                String::from(content),
+                                &HashMap::new(),
+                            )?;
+
+                            let is_formatted = pretty != content;
+
+                            if is_formatted
+                                && name.eq_ignore_ascii_case("template")
+                                && config.indent_template
+                            {
+                                pretty.replace('\n', {
+                                    let width = usize::from(config.indent_width);
+                                    let indent = if config.use_tabs {
+                                        format!("{:\t<width$}", "")
+                                    } else {
+                                        format!("{: <width$}", "")
+                                    };
+
+                                    &format!("\n{}", indent)
+                                })
+                            } else {
+                                pretty
+                            }
+                        };
+
+                        buffer.push_str(&pretty);
                     } else {
                         buffer.push_str(content);
                     }