+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added configuration:
+
+  | Key              | Default | Description                                |
+  | ---------------- | ------- | ------------------------------------------ |
+  | `indentTemplate` | `true`  | Indent the content of the `<template>` tag |
+  | `indentWidth`    | `2`     | Width of the indentation                   |
+  | `useTabs`        | `false` | Use tabs for indentation                   |
+
 ## [0.2.0] - 2021-01-25
 
 ### Added
@@ -42,4 +52,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.2.0]: https://github.com/malobre/dprint-plugin-vue/compare/v0.1.0...v0.2.0
 [0.1.1]: https://github.com/malobre/dprint-plugin-vue/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/malobre/dprint-plugin-vue/releases/tag/v0.1.0
-